Fix UiConfig test
