+++ zipkin-lens/package.json
@@ -52,7 +52,6 @@
     "eslint-plugin-react": "^7.12.4",
     "eslint-plugin-react-hooks": "^2.5.0",
     "events": "^3.3.0",
-    "fetch-mock": "^9.4.0",
     "history": "^4.10.1",
     "http-proxy-middleware": "^1.0.3",
     "i18next": "^23.8.2",
